# internal/apiserver/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  core.ErrResponse:
    properties:
      code:
        description: Code 定义了业务错误代码
        type: integer
      message:
        description: |-
          Message 包含此消息的详细信息
          此消息适合暴露给外部
        type: string
      reference:
        description: Reference 返回参考文档，可能有助于解决此错误
        type: string
    type: object
  core.Response:
    properties:
      code:
        description: Code 业务状态码，0 表示成功
        type: integer
      data:
        description: Data 响应数据
      message:
        description: Message 响应消息
        type: string
      reference:
        description: Reference 返回参考文档（错误时使用）
        type: string
    type: object
...
  request.UpdateTesteeRequest:
    properties:
      birthday:
        description: 出生日期
        type: string
      gender:
        description: 性别
        type: string
      is_key_focus:
        description: 是否重点关注
        type: boolean
      name:
        description: 姓名
        type: string
    type: object
...
  response.AssessmentStatsResponse:
    properties:
      last_assessment_at:
        description: 最后测评时间
        type: string
      last_risk_level:
        description: 最后风险等级
        type: string
      last_risk_level_label:
        description: 最后风险等级中文
        type: string
      total_count:
        description: 总次数
        type: integer
    type: object
...
  response.GuardianResponse:
    properties:
      name:
        type: string
      phone:
        type: string
      relation:
        type: string
    type: object
...
  response.TesteeListResponse:
    properties:
      items:
        description: 列表数据
        items:
          $ref: '#/definitions/response.TesteeResponse'
        type: array
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  response.TesteeResponse:
    properties:
      assessment_stats:
        allOf:
        - $ref: '#/definitions/response.AssessmentStatsResponse'
        description: 测评统计
      birthday:
        description: 出生日期
        type: string
      created_at:
        description: 创建时间
        type: string
      gender:
        description: 性别
        type: string
      gender_label:
        description: 性别中文
        type: string
      guardians:
        description: 监护人信息列表
        items:
          $ref: '#/definitions/response.GuardianResponse'
        type: array
      id:
        description: ID
        type: string
      is_key_focus:
        description: 是否重点关注
        type: boolean
      is_key_focus_label:
        description: 是否重点关注中文
        type: string
      name:
        description: 姓名
        type: string
      org_id:
        description: 机构ID
        type: string
      profile_id:
        description: 用户档案ID
        type: string
      source:
        description: 来源
        type: string
      source_label:
        description: 来源中文
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: Questionnaire Scale API server (REST & gRPC)
  title: QS API Server
  version: "1.0"
paths:
...
  /api/v1/testees:
    get:
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TesteeListResponse'
              type: object
      summary: 查询受试者列表
      tags:
      - 受试者
  /api/v1/testees/{id}:
    get:
      description: 根据受试者 ID 获取受试者详情
      parameters:
      - description: 受试者ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TesteeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrResponse'
      security:
      - BearerAuth: []
      summary: 获取受试者详情
      tags:
      - 受试者
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 受试者ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新受试者请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTesteeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TesteeResponse'
              type: object
      summary: 更新受试者
      tags:
      - 受试者
...
  /api/v1/testees/{id}/scale-analysis:
    get:
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 受试者ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Response'
      summary: 获取受试者量表分析
      tags:
      - 受试者
...
  /api/v1/testees/by-profile-id:
    get:
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: Profile ID
        in: query
        name: profile_id
        required: true
        type: string
      - description: 机构 ID
        in: query
        name: org_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/core.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.TesteeResponse'
              type: object
      summary: 根据 profile_id 获取受试者
      tags:
      - 受试者
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"